requirements.txt
asgiref==3.3.1
certifi==2020.12.5
chardet==4.0.0
cloudinary==1.25.0
Django==3.1.7
django-cloudinary-storage==0.3.0
django-cors-headers==3.7.0
djangorestframework==3.12.2
djangorestframework-camel-case==1.2.0
idna==2.10
Pillow==8.1.2
psycopg2==2.8.6
python-decouple==3.4
pytz==2021.1
requests==2.25.1
six==1.15.0
sqlparse==0.4.1
urllib3==1.26.4
